Fix test abous RegExp evaluation (ES2018).
There was a typo in RegExp look-behind, so the test was always false.

Use `new RegExp` instead of `eval` to limit the eval usage.

The regexp-lookbehind feature was renamed to regexp-look-behind which is
more correct. It is a breaking change but it should be ok because the
project is still young.